Damage follows level of difficulty

+++ Assets/Scripts/Scriptable Objects/GunData.cs
@@ -9,7 +9,7 @@ public class GunData : ScriptableObject
     public new string name;
 
     [Header("Shooting")]
-    public float damage;
+    public int damage;
     public float maxDistance;
 
     [Header("Reloading")]
@@ -19,4 +19,5 @@ public class GunData : ScriptableObject
     public float reloadTime;
     [HideInInspector]
     public bool isReloading;
+    
 }
